+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Fixed a bug when displaying a request failure that includes a localized
   message in the error details (#293).
+* gargle is better able to detect when it's running on Posit Workbench, but not
+  necessarily in RStudio, such as in Positron or VS Code (#291).
 
 # gargle 1.5.2